+++ packages/cli/src/entrypoint.ts
@@ -1,9 +1,6 @@
 import nodePath from "path";
 import { Item } from "./item";
-import resolve from "resolve";
-import { EXTENSIONS } from "./constants";
 import { Package } from "./package";
-import { FatalError } from "./errors";
 import { JSONValue } from "./utils";
 import normalizePath from "normalize-path";
 
@@ -18,9 +15,16 @@ export class Entrypoint extends Item<{
   };
 }> {
   package: Package;
-  constructor(filePath: string, contents: string, pkg: Package) {
-    super(filePath, contents);
+  source: string;
+  constructor(
+    filePath: string,
+    contents: string,
+    pkg: Package,
+    source: string
+  ) {
+    super(filePath, contents, pkg._jsonDataByPath);
     this.package = pkg;
+    this.source = source;
   }
 
   get name(): string {
@@ -30,53 +34,5 @@ export class Entrypoint extends Item<{
         nodePath.relative(this.package.directory, this.directory)
       )
     );
-  }
-
-  get configSource(): string {
-    if (this.package.project.experimentalFlags.newEntrypoints) {
-      if (this.json.preconstruct.source !== undefined) {
-        throw new FatalError(
-          "The source option is not allowed with the newEntrypoints experimental flag",
-          this.name
-        );
-      }
-      return nodePath.relative(
-        this.directory,
-        nodePath.join(
-          this.package.directory,
-          "src",
-          nodePath.relative(this.package.directory, this.directory)
-        )
-      );
-    }
-    if (
-      this.json.preconstruct.source !== undefined &&
-      typeof this.json.preconstruct.source !== "string"
-    ) {
-      throw new FatalError(
-        "The source option for this entrypoint is not a string",
-        this.name
-      );
-    }
-    if (this.json.preconstruct.source === undefined) {
-      return "src/index";
-    }
-    return this.json.preconstruct.source;
-  }
-  _sourceCached?: string;
-  get source(): string {
-    if (this._sourceCached === undefined) {
-      this._sourceCached = resolve.sync(
-        nodePath.join(this.directory, this.configSource),
-        {
-          extensions: EXTENSIONS,
-        }
-      );
-    }
-    return this._sourceCached;
-  }
-  updater(json: typeof Entrypoint.prototype.json) {
-    super.updater(json);
-    this._sourceCached = undefined;
   }
 }