+++ packages/cli/src/utils.ts
@@ -64,7 +64,7 @@ export function getEntrypointName(pkg: Package, entrypointDir: string) {
   );
 }
 
-type DistFilenameStrategy = "full" | "only-unscoped-package-name";
+type DistFilenameStrategy = "full" | "unscoped-package-name";
 
 function getDistNameWithStrategy(
   pkg: Package,
@@ -85,38 +85,31 @@ function getDistName(
   if (forceStrategy) {
     return getDistNameWithStrategy(pkg, entrypointName, forceStrategy);
   }
-  if (pkg.project.experimentalFlags.newDistFilenames) {
-    if ("distFilenameStrategy" in pkg.project.json.preconstruct) {
-      if (
-        pkg.project.json.preconstruct.distFilenameStrategy !== "full" &&
-        pkg.project.json.preconstruct.distFilenameStrategy !==
-          "only-unscoped-package-name"
-      ) {
-        throw new FatalError(
-          `distFilenameStrategy is defined in your Preconstruct config as ${JSON.stringify(
-            pkg.project.json.preconstruct.distFilenameStrategy
-          )} but the only accepted values are "full" and "only-unscoped-package-name"`,
-          pkg.project.name
-        );
-      }
-      if (
-        pkg.project.json.preconstruct.distFilenameStrategy ===
-        "only-unscoped-package-name"
-      ) {
-        return getDistNameWithStrategy(
-          pkg,
-          entrypointName,
-          "only-unscoped-package-name"
-        );
-      }
+  if ("distFilenameStrategy" in pkg.project.json.preconstruct) {
+    if (
+      pkg.project.json.preconstruct.distFilenameStrategy !== "full" &&
+      pkg.project.json.preconstruct.distFilenameStrategy !==
+        "unscoped-package-name"
+    ) {
+      throw new FatalError(
+        `distFilenameStrategy is defined in your Preconstruct config as ${JSON.stringify(
+          pkg.project.json.preconstruct.distFilenameStrategy
+        )} but the only accepted values are "full" and "unscoped-package-name"`,
+        pkg.project.name
+      );
     }
-    return getDistNameWithStrategy(pkg, entrypointName, "full");
+    if (
+      pkg.project.json.preconstruct.distFilenameStrategy ===
+      "unscoped-package-name"
+    ) {
+      return getDistNameWithStrategy(
+        pkg,
+        entrypointName,
+        "unscoped-package-name"
+      );
+    }
   }
-  return getDistNameWithStrategy(
-    pkg,
-    entrypointName,
-    "only-unscoped-package-name"
-  );
+  return getDistNameWithStrategy(pkg, entrypointName, "full");
 }
 
 export const validFieldsFromPkg = {